Fix code for Acceptance Tests

See merge request se-2022/group-23/EzWh!31

+++ code/server/api/components/user/controller.js
@@ -95,22 +95,36 @@ class UserController {
 	}
 
 	async modifyRight(username, oldType, newType) {
-		if (oldType === User.MANAGER || oldType === User.ADMINISTRATOR)
-			throw UserErrorFactory.newAttemptCreationPrivilegedAccount();
+		try {
+			if (!User.isValidType(oldType) || !User.isValidType(newType))
+				throw UserErrorFactory.newTypeNotFound422();
 
-		const user = await this.dao.getUserByEmailAndType(username, oldType);
-		if (user === undefined)
-			throw UserErrorFactory.newUserNotFound();
+			if (oldType === User.MANAGER || oldType === User.ADMINISTRATOR)
+				throw UserErrorFactory.newAttemptCreationPrivilegedAccount();
 
-		const { changes } = await this.dao.modifyRight(username, oldType, newType);
-		if (changes === 0)
-			throw UserErrorFactory.newAttemptCreationPrivilegedAccount();
+			const user = await this.dao.getUserByEmailAndType(username, oldType);
+			if (user === undefined)
+				throw UserErrorFactory.newUserNotFound();
 
+			const { changes } = await this.dao.modifyRight(username, oldType, newType);
+			if (changes === 0)
+				throw UserErrorFactory.newAttemptCreationPrivilegedAccount();
+		} catch (err) {
+			if (err.code === "SQLITE_CONSTRAINT") {
+				if (err.message.includes("user.email, user.type"))
+					err = UserErrorFactory.newUserConflict();
+			}
+
+			throw err;
+		}
 	}
 
 	async deleteUser(username, type) {
 		if (type === User.MANAGER || type === User.ADMINISTRATOR)
 			throw UserErrorFactory.newAttemptCreationPrivilegedAccount();
+
+		if (!User.isValidType(type))
+			throw UserErrorFactory.newTypeNotFound422();
 
 		await this.dao.deleteUser(username, type);
 	}
